Dev: Get up to date (admin, videos, user managment)

+++ server/controllers/users.controller.js
@@ -3,7 +3,12 @@ import { User } from "../models/user.model.js";
 export const getUsers = async (req, res) => {
     // get all the existing users
     try {
-        const users = await User.find();
+        const requester = await User.findById(req.user.id);
+        if (requester.role !== "admin") {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
+        const users = await User.find({ role: "user" });
         res.json(users);
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -13,6 +18,11 @@ export const getUsers = async (req, res) => {
 export const getUser = async (req, res) => {
     // get user by id
     try {
+        const requester = await User.findById(req.user.id);
+        if (requester.role !== "admin") {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
         const user = await User.findById(req.params.id);
         res.json(user);
     } catch (error) {
@@ -23,6 +33,11 @@ export const getUser = async (req, res) => {
 export const updateUser = async (req, res) => {
     // update user by id
     try {
+        const requester = await User.findById(req.user.id);
+        if (requester.role !== "admin") {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
         const { username, email, password, role } = req.body;
 
         const noteUpdated = await User.findOneAndUpdate(
@@ -44,8 +59,32 @@ export const updateUser = async (req, res) => {
 export const deleteUser = async (req, res) => {
     // delete user by id
     try {
+        const requester = await User.findById(req.user.id);
+        if (requester.role !== "admin") {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
         const userDeleted = await User.findByIdAndDelete(req.params.id);
         return res.json(userDeleted);
+    } catch (error) {
+        return res.status(500).json({ message: error.message });
+    }
+};
+
+export const makeAdmin = async (req, res) => {
+    // make user admin by id
+    try {
+        const requester = await User.findById(req.user.id);
+
+        if (requester.role !== "admin") {
+            return res.status(401).json({ message: "Unauthorized" });
+        }
+
+        const user = await User.findById(req.params.id);
+        user.role = "admin";
+
+        await user.save();
+        return res.json(user);
     } catch (error) {
         return res.status(500).json({ message: error.message });
     }
